+++ inst/examples/shinylive/dashboard/app.R
@@ -1,3 +1,5 @@
+webr::install("tablerDash", repos = c("https://rinterface.github.io/rinterface-wasm-cran/", "https://repo.r-wasm.org"))
+
 library(shiny)
 library(tablerDash)
 library(echarts4r)
@@ -107,7 +109,7 @@ shiny::shinyApp(
     title = "tablerDash",
     body = tablerDashBody(
 
-      chooseSliderSkin("Nice"),
+      chooseSliderSkin("Modern"),
       tablerTabItems(
         tablerTabItem(
           tabName = "Home",