+++ NEWS.md
@@ -1,5 +1,9 @@
 # tablerDash 0.1.5.9000
 
+## Documentation
+- Upgrade pkgdown setup.
+- Added shinylive support.
+
 ## Bug Fix
 - Resizing a `tablerCard()` triggers the shown javascript event, telling shiny to 
 update the size of the contained object.